fix(advisor): launch local review from trusted code
